third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/batch_normalization.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] batchNormalization float32 2D constant tensors default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test batchNormalization float32 actual -4.312742233276367 should be close enough to expected -6.995829105377197 by ULP distance: expected a number less than or equal to 6n but got 5626841n
[FAIL] batchNormalization float16 1D tensor options.axis=0
  assert_less_than_equal: assert_array_approx_equals_ulp: test batchNormalization float16 actual 0 should be close enough to expected -4.3125 by ULP distance: expected a number less than or equal to 6 but got 50256
[FAIL] batchNormalization float16 2D tensor (mean and variance are non-constant) default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test batchNormalization float16 actual 0 should be close enough to expected -4.3125 by ULP distance: expected a number less than or equal to 6 but got 50256
[FAIL] batchNormalization float16 2D constant tensors default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test batchNormalization float16 actual 0 should be close enough to expected -4.3125 by ULP distance: expected a number less than or equal to 6 but got 50256
[FAIL] batchNormalization float16 2D tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test batchNormalization float16 actual 0 should be close enough to expected -4.3125 by ULP distance: expected a number less than or equal to 6 but got 50256
[FAIL] batchNormalization float16 3D tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test batchNormalization float16 actual 0.0078125 should be close enough to expected -12.6328125 by ULP distance: expected a number less than or equal to 6 but got 43601
[FAIL] batchNormalization float16 4D tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test batchNormalization float16 actual -2 should be close enough to expected -16.84375 by ULP distance: expected a number less than or equal to 6 but got 3126
[FAIL] batchNormalization float16 4D NCHW tensor options.axis=1
  assert_less_than_equal: assert_array_approx_equals_ulp: test batchNormalization float16 actual -2 should be close enough to expected -16.84375 by ULP distance: expected a number less than or equal to 6 but got 3126
[FAIL] batchNormalization float16 4D NHWC tensor options.axis=3
  assert_less_than_equal: assert_array_approx_equals_ulp: test batchNormalization float16 actual -2 should be close enough to expected -16.84375 by ULP distance: expected a number less than or equal to 6 but got 3126
[FAIL] batchNormalization float16 4D NCHW tensor options.scale
  assert_less_than_equal: assert_array_approx_equals_ulp: test batchNormalization float16 actual -512 should be close enough to expected -1103 by ULP distance: expected a number less than or equal to 6 but got 1103
[FAIL] batchNormalization float16 4D NCHW tensor options.bias
  assert_less_than_equal: assert_array_approx_equals_ulp: test batchNormalization float16 actual 512 should be close enough to expected 47.34375 by ULP distance: expected a number less than or equal to 6 but got 3605
[FAIL] batchNormalization float16 4D NCHW tensor options.epsilon
  assert_less_than_equal: assert_array_approx_equals_ulp: test batchNormalization float16 actual -2 should be close enough to expected -16.84375 by ULP distance: expected a number less than or equal to 6 but got 3126
[FAIL] batchNormalization float16 4D NHWC tensor all options
  assert_less_than_equal: assert_array_approx_equals_ulp: test batchNormalization float16 actual -512 should be close enough to expected -1039 by ULP distance: expected a number less than or equal to 6 but got 1039
Harness: the test ran to completion.
